+++ TagCloud/Interfaces/IProgramInterface.cs
@@ -1,0 +1,7 @@
+﻿namespace TagCloud.Interfaces
+{
+    public interface IProgramInterface
+    {
+        void Run(string[] args);
+    }
+}
